+++ src/Structs/dynarr.h
@@ -1,7 +1,7 @@
 #ifndef CLIBS_DYNAMIC_ARRAY_H
 #define CLIBS_DYNAMIC_ARRAY_H
 
-#include "../array_printf.h"   /* for list_printf */
+#include "../Dev/array_printf.h"
 #include "../Dev/attributes.h" /* Private */
 
 #include <stdbool.h>
